Use fetch rather than axios for requests from the client
Note there are no particular technical reasons or personal preferences
for making this change, either is a reasonable decision.

The CYF syllabus uses fetch, so this reduces new context for students:
https://syllabus.codeyourfuture.io/react/week-3/lesson

Inlining the fetch into the component isn't good testing or
architectural practice, as it couples the component to the fetch API
which we don't own, but again reduces new context for CYF students.